Fix Tetranacci_Number_II to use long long int for array and add Make It problem statement for HackerRank final exam

+++ CS/Introduction to algorithm/final exam/Tetranacci_Number_II.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main(){
     int n;
     cin >> n;
-    int treta[n+1];
+    long long int treta[n+1];
     treta[0] = 0;
     treta[1] = 1;
     treta[2] = 1;
